Factored out a separate file for class for AquariusFileHandler
